modified structs wrr_rq and sched_wrr_entity and added prototype declaration of wrr_periodic_load_balance

+++ include/linux/sched.h
@@ -500,10 +500,8 @@ struct sched_rt_entity {
 struct sched_wrr_entity {
 	struct list_head	run_list;
 	unsigned int		wrr_se_timeslice;
-	unsigned long		wrr_se_weight;
+	int			wrr_se_weight;
 	unsigned long		timeout;
-	//struct sched_wrr_entity	*parent;
-	//struct wrr_rq		*wrr_rq;
 };
 
 struct sched_dl_entity {
